requirements.txt
grpcio>=1.54.0
grpcio-tools>=1.54.0
protobuf>=4.20.0
pytest>=7.0.0
pytest-asyncio>=0.22.0
allure-pytest>=2.11.0